requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
celery==5.2.7
redis==4.3.4
django-celery-beat==2.2.1
django-celery-results==2.3.0
python-dotenv==0.21.0
gunicorn==20.1.0
django-environ==0.8.1
reportlab==3.6.2
